# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YTTracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Channel {
  id: number;
  name: string;
  youtubeChannelId: string;
  thumbnailUrl?: string | null;
  createdAt: string;
  videoCount: number;
  pendingCount: number;
}

export interface CreateChannelRequest {
  name: string;
  youtubeChannelId: string;
  thumbnailUrl?: string | null;
}

export type VideoStatus = (typeof VideoStatus)[keyof typeof VideoStatus];

export const VideoStatus = {
  pending: "pending",
  watched: "watched",
  skipped: "skipped",
} as const;

export interface Video {
  id: number;
  channelId: number;
  channelName: string;
  youtubeVideoId: string;
  title: string;
  thumbnailUrl?: string | null;
  durationSeconds: number;
  status: VideoStatus;
  publishedAt: string;
  watchedAt?: string | null;
  createdAt: string;
}

export interface CreateVideoRequest {
  channelId: number;
  youtubeVideoId: string;
  title: string;
  thumbnailUrl?: string | null;
  durationSeconds: number;
  publishedAt: string;
}

export type UpdateVideoStatusRequestStatus =
  (typeof UpdateVideoStatusRequestStatus)[keyof typeof UpdateVideoStatusRequestStatus];

export const UpdateVideoStatusRequestStatus = {
  pending: "pending",
  watched: "watched",
  skipped: "skipped",
} as const;

export interface UpdateVideoStatusRequest {
  status: UpdateVideoStatusRequestStatus;
}

export interface DailyStats {
  date: string;
  watchedSeconds: number;
  watchedCount: number;
  skippedCount: number;
}

export interface DailyStatsResponse {
  days: DailyStats[];
  totalPendingSeconds: number;
  totalPendingCount: number;
  averageDailyWatchSeconds: number;
}

export interface ChannelInsight {
  channelId: number;
  channelName: string;
  thumbnailUrl?: string | null;
  totalVideos: number;
  watchedVideos: number;
  skippedVideos: number;
  pendingVideos: number;
  watchRatio: number;
  avgDaysToWatch?: number | null;
  totalWatchedSeconds: number;
  recommendUnsubscribe: boolean;
}

export interface InsightsResponse {
  channels: ChannelInsight[];
  topChannels: ChannelInsight[];
  unsubscribeRecommendations: ChannelInsight[];
}

export interface SmartPlaylistRequest {
  /** Available time in minutes for this session */
  sessionMinutes: number;
  prioritizeTopChannels?: boolean;
  /** Optional filter to specific channels */
  channelIds?: number[];
}

export interface SmartPlaylistResponse {
  videos: Video[];
  totalDurationSeconds: number;
  sessionMinutes: number;
}

export type ListVideosParams = {
  channelId?: number;
  status?: ListVideosStatus;
};

export type ListVideosStatus =
  (typeof ListVideosStatus)[keyof typeof ListVideosStatus];

export const ListVideosStatus = {
  pending: "pending",
  watched: "watched",
  skipped: "skipped",
} as const;

export type GetDailyStatsParams = {
  days?: number;
};