# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YTTracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Channel {
  id: number;
  name: string;
  youtubeChannelId: string;
  thumbnailUrl?: string | null;
  createdAt: Date;
  videoCount: number;
  pendingCount: number;
}