# lib/api-zod/src/generated/types/createChannelRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YTTracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateChannelRequest {
  name: string;
  youtubeChannelId: string;
  thumbnailUrl?: string | null;
}