# lib/api-zod/src/generated/types/createVideoRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YTTracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVideoRequest {
  channelId: number;
  youtubeVideoId: string;
  title: string;
  thumbnailUrl?: string | null;
  durationSeconds: number;
  publishedAt: Date;
}