# lib/api-zod/src/generated/types/dailyStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YTTracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyStats {
  date: string;
  watchedSeconds: number;
  watchedCount: number;
  skippedCount: number;
}