# lib/api-zod/src/generated/types/dailyStatsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YTTracker API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyStats } from "./dailyStats";

export interface DailyStatsResponse {
  days: DailyStats[];
  totalPendingSeconds: number;
  totalPendingCount: number;
  averageDailyWatchSeconds: number;
}