# lib/api-zod/src/generated/types/insightsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YTTracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelInsight } from "./channelInsight";

export interface InsightsResponse {
  channels: ChannelInsight[];
  topChannels: ChannelInsight[];
  unsubscribeRecommendations: ChannelInsight[];
}