# lib/api-zod/src/generated/types/smartPlaylistRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YTTracker API
 * OpenAPI spec version: 0.1.0
 */

export interface SmartPlaylistRequest {
  /** Available time in minutes for this session */
  sessionMinutes: number;
  prioritizeTopChannels?: boolean;
  /** Optional filter to specific channels */
  channelIds?: number[];
}