# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YTTracker API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoStatus } from "./videoStatus";

export interface Video {
  id: number;
  channelId: number;
  channelName: string;
  youtubeVideoId: string;
  title: string;
  thumbnailUrl?: string | null;
  durationSeconds: number;
  status: VideoStatus;
  publishedAt: Date;
  watchedAt?: Date | null;
  createdAt: Date;
}